added custom text option and dark mode
added the ability for users to add their own text to be read back, for those instances where information is not in pdf form. added a dark mode for those who do not like being blinded by white backgrounds. fixed an issue with the pause button not being clickable after being used.